setUp add tearDown work

+++ src/main/java/org/example/models/DefaultFields.java
@@ -1,5 +1,5 @@
 package org.example.models;
 
 public abstract class DefaultFields {
-   abstract public short getId();
+   abstract public int getId();
 }
