+++ src/main/java/org/example/models/Fabricator.java
@@ -3,12 +3,14 @@ package org.example.models;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.ToString;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
 @AllArgsConstructor
 public class Fabricator extends DefaultFields {
-  short id;
+  int id;
   String name;
   String country;
+
 }
